requirements.txt
requests==2.24.0
selenium==3.141.0
pandas==1.3.3
pathlib2==2.3.5
keyring==22.3.0
pyyaml==5.4.1
bs4==0.0.1
black==21.12b0
SQLAlchemy==1.4.3
aiosqlite==0.17.0
kucoin-python==1.0.11
Sphinx==4.4.0
sphinx-rtd-theme==1.0.0
recommonmark==0.7.1
networkx==2.7.1